ISPN-3435 Enable tests for RemoteCache.getBulk() and keySet()
