+++ client/hotrod-client/src/test/java/org/infinispan/client/hotrod/BaseBulkGetKeysTest.java
@@ -14,6 +14,7 @@ import org.infinispan.manager.EmbeddedCacheManager;
 import org.infinispan.server.hotrod.HotRodServer;
 import org.infinispan.test.MultipleCacheManagersTest;
 import org.testng.annotations.AfterClass;
+import org.testng.annotations.Test;
 
 /**
  * Tests functionality related to getting multiple entries from a HotRod server
@@ -22,6 +23,7 @@ import org.testng.annotations.AfterClass;
  * @author <a href="mailto:rtsang@redhat.com">Ray Tsang</a>
  * @since 5.2
  */
+@Test(groups = "functional")
 public abstract class BaseBulkGetKeysTest extends MultipleCacheManagersTest {
 	protected HotRodServer[] hotrodServers;
 	protected RemoteCacheManager remoteCacheManager;
@@ -69,7 +71,7 @@ public abstract class BaseBulkGetKeysTest extends MultipleCacheManagersTest {
 			assert set.contains(i);
 		}
 	}
-	
+
 	public void testBulkGetAfterLifespanExpire() throws InterruptedException {
 		Map<String, String> dataIn = new HashMap<String, String>();
 		dataIn.put("aKey", "aValue");
